+++ src/components/pages/apps/genes/gsea/gsea-plot/pathway-filter.tsx
@@ -1,0 +1,95 @@
+import { CheckPropRow } from '@/components/dialogs/check-prop-row'
+import { NumericalInput } from '@/components/shadcn/ui/themed/numerical-input'
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from '@/components/shadcn/ui/themed/v2/popover'
+import { ToolbarIconButton } from '@/components/toolbar/toolbar-icon-button'
+import { produce } from 'immer'
+import { ListFilter } from 'lucide-react'
+import { useState } from 'react'
+import { useGseaSettings } from './gsea-settings-store'
+
+export function PathwayFilter() {
+  const { settings, updateSettings } = useGseaSettings()
+
+  const [open, setOpen] = useState(false)
+
+  return (
+    <Popover open={open} onOpenChange={setOpen}>
+      <PopoverTrigger
+        render={
+          <ToolbarIconButton checked={open} title="Filter Pathways">
+            <ListFilter size={14} />
+          </ToolbarIconButton>
+        }
+      />
+
+      <PopoverContent
+        //onEscapeKeyDown={() => setOpen(false)}
+        //onInteractOutside={() => setOpen(false)}
+        //align="end"
+        className="px-3 py-3 gap-y-2"
+      >
+        <h2 className="font-bold">Pathway Filters</h2>
+        <CheckPropRow
+          title={<span>NES &ge;</span>}
+          checked={settings.filters.nes.on}
+          onCheckedChange={(v) => {
+            updateSettings(
+              produce(settings, (draft) => {
+                draft.filters.nes.on = v
+              })
+            )
+          }}
+        >
+          <NumericalInput
+            value={settings.filters.nes.value}
+            h="md"
+            placeholder="NES"
+            limit={[0, 100]}
+            step={0.01}
+            dp={2}
+            onNumChanged={(v) => {
+              updateSettings(
+                produce(settings, (draft) => {
+                  draft.filters.nes.value = v
+                })
+              )
+            }}
+            w="sm"
+          />
+        </CheckPropRow>
+        <CheckPropRow
+          title={<span>FDR q-value &le;</span>}
+          checked={settings.filters.q.on}
+          onCheckedChange={(v) => {
+            updateSettings(
+              produce(settings, (draft) => {
+                draft.filters.q.on = v
+              })
+            )
+          }}
+        >
+          <NumericalInput
+            value={settings.filters.q.value}
+            h="md"
+            placeholder="FDR q-value"
+            limit={[0, 1]}
+            step={0.0001}
+            dp={4}
+            onNumChanged={(v) => {
+              updateSettings(
+                produce(settings, (draft) => {
+                  draft.filters.q.value = v
+                })
+              )
+            }}
+            w="sm"
+          />
+        </CheckPropRow>
+      </PopoverContent>
+    </Popover>
+  )
+}