+++ src/components/pages/apps/genes/gsea/gsea-plot/toolbars/home.tsx
@@ -36,6 +36,7 @@ export function HomeToolbar() {
                   if (!success) {
                     return
                   }
+
                   loadGseaZipWithErrorHandling(files)
                 })
               },